ci: add cargo test and clippy workflow
GitHub Actions runs build, test, and clippy on push/PR to main.
Suppress clippy on generated protobuf code with #[allow(clippy::all)].
Fix unnecessary cast in test code.

+++ src/models.rs
@@ -205,7 +205,7 @@ mod tests {
             trains.push(Train {
                 route: format!("{}", i + 1),
                 destination: format!("Dest {}", i),
-                minutes: i as i32,
+                minutes: i,
                 is_express: false,
                 arrival_timestamp: 0.0,
                 direction: Direction::Uptown,
